refactor: 컨트롤러 상속 제거
